+++ backend/agile-organizer/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/model/BaseEntity.java
@@ -1,0 +1,28 @@
+import javax.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+import java.io.Serializable;
+import java.time.Instant;
+
+@MappedSuperclass
+@Data
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+public class BaseEntity implements Serializable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY) 
+    private int id;
+
+    @Builder.Default
+    private Boolean deleted = false;
+
+    @Column(name = "created_at", nullable = false, updatable = false)
+    @Builder.Default
+    private Instant createdAt = Instant.now();
+}
