+++ backend/agile-organizer/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/model/State.java
@@ -1,0 +1,5 @@
+public class State extends BaseEntity{
+
+
+    
+}
