+++ backend/agile-organizer/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/model/ToDoItem.java
@@ -1,74 +1,25 @@
 package com.springboot.MyTodoList.model;
 
-
 import javax.persistence.*;
 import java.time.OffsetDateTime;
 
-/*
-    representation of the TODOITEM table that exists already
-    in the autonomous database
- */
 @Entity
-@Table(name = "TODOITEM")
-public class ToDoItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int ID;
-    @Column(name = "DESCRIPTION")
-    String description;
-    @Column(name = "CREATION_TS")
-    OffsetDateTime creation_ts;
-    @Column(name = "done")
-    boolean done;
-    public ToDoItem(){
+@Table(name = "TODO")
+public class ToDoItem extends BaseEntity{
 
-    }
-    public ToDoItem(int ID, String description, OffsetDateTime creation_ts, boolean done) {
-        this.ID = ID;
-        this.description = description;
-        this.creation_ts = creation_ts;
-        this.done = done;
-    }
+    private String title;
+    private String description;
 
-    public int getID() {
-        return ID;
-    }
+    private State state;
+    private Sprint sprint;
 
-    public void setID(int ID) {
-        this.ID = ID;
-    }
+    private User user;
+    private Project project;
+    private int storyPoints;
 
-    public String getDescription() {
-        return description;
-    }
+    private OffsetDateTime dueDate;
 
-    public void setDescription(String description) {
-        this.description = description;
-    }
+    private float estimatedHours;
+    private float realHours;
 
-    public OffsetDateTime getCreation_ts() {
-        return creation_ts;
-    }
-
-    public void setCreation_ts(OffsetDateTime creation_ts) {
-        this.creation_ts = creation_ts;
-    }
-
-    public boolean isDone() {
-        return done;
-    }
-
-    public void setDone(boolean done) {
-        this.done = done;
-    }
-
-    @Override
-    public String toString() {
-        return "ToDoItem{" +
-                "ID=" + ID +
-                ", description='" + description + '\'' +
-                ", creation_ts=" + creation_ts +
-                ", done=" + done +
-                '}';
-    }
 }
